Refactor HNSW implementation and add comprehensive tests
- Implemented HNSW search and insertion logic in hnsw_query.go
- Added unit tests for distance calculations, heap operations, and HNSW core functionalities in hnsw_test.go
- Introduced utility functions for managing HNSW layers and neighbors in hnsw_utils.go
- Enhanced performance benchmarks for insertion and search operations
- Included persistence tests to ensure data integrity after saving and loading collections
- Conducted robustness and concurrency tests to validate system stability under load

+++ kernel/vectordb/vamana/index.go
@@ -137,10 +137,10 @@ func (idx *VamanaIndex) NumPoints() uint64 {
 }
 
 // NumPointsTotal 返回索引中的总点数 (含已删除)
-func (idx *VamanaIndex) NumPointsTotal() int {
+func (idx *VamanaIndex) NumPointsTotal() uint64 {
 	idx.mu.RLock()
 	defer idx.mu.RUnlock()
-	return len(idx.vectors)
+	return uint64(len(idx.vectors))
 }
 
 // Close 释放索引关联的资源。
